[ADD/#20] shared/lib/hooks에 공용 fade slide 애니메이션 스크립트 추가
