+++ src/widgets/chat/bottom-snackbar/ui/BottomSnackbar.tsx
@@ -1,11 +1,10 @@
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Animated, Easing, Text, View } from "react-native";
+import { Animated, Text, View } from "react-native";
+import { useFadeSlideVisibility } from "@/shared/lib/hooks/useFadeSlideVisibility";
 import { shadows } from "@/shared/styles/shadows";
 import type { BottomSnackbarProps } from "../model/types";
 
 const DEFAULT_HORIZONTAL_INSET = 24;
 const DEFAULT_BOTTOM_OFFSET = 0;
-const ANIMATION_DURATION_MS = 180;
 
 export function BottomSnackbar({
 	visible,
@@ -16,36 +15,7 @@ export function BottomSnackbar({
 	horizontalInset = DEFAULT_HORIZONTAL_INSET,
 	testID,
 }: BottomSnackbarProps) {
-	const [shouldRender, setShouldRender] = useState(visible);
-	const progress = useRef(new Animated.Value(visible ? 1 : 0)).current;
-
-	useEffect(() => {
-		if (visible) {
-			setShouldRender(true);
-		}
-
-		Animated.timing(progress, {
-			toValue: visible ? 1 : 0,
-			duration: ANIMATION_DURATION_MS,
-			easing: Easing.out(Easing.cubic),
-			useNativeDriver: true,
-		}).start(({ finished }) => {
-			if (!finished) return;
-			if (!visible) setShouldRender(false);
-		});
-	}, [progress, visible]);
-
-	const animatedStyle = useMemo(() => {
-		const translateY = progress.interpolate({
-			inputRange: [0, 1],
-			outputRange: [10, 0],
-		});
-
-		return {
-			opacity: progress,
-			transform: [{ translateY }],
-		} as const;
-	}, [progress]);
+	const { shouldRender, animatedStyle } = useFadeSlideVisibility(visible);
 
 	if (!shouldRender) return null;
